+++ Userland/Libraries/LibWeb/CSS/StyleValue.h
@@ -267,6 +267,8 @@ private:
 
 using StyleValueVector = Vector<ValueComparingNonnullRefPtr<StyleValue const>>;
 
+Optional<CSS::Length> absolutized_length(CSS::Length const&, CSSPixelRect const& viewport_rect, Gfx::FontPixelMetrics const&, CSSPixels font_size, CSSPixels root_font_size, CSSPixels line_height, CSSPixels root_line_height);
+
 class StyleValue : public RefCounted<StyleValue> {
 public:
     virtual ~StyleValue() = default;
@@ -500,39 +502,6 @@ struct StyleValueWithDefaultOperators : public StyleValue {
         auto const& typed_other = static_cast<T const&>(other);
         return static_cast<T const&>(*this).properties_equal(typed_other);
     }
-};
-
-class BorderRadiusStyleValue final : public StyleValueWithDefaultOperators<BorderRadiusStyleValue> {
-public:
-    static ValueComparingNonnullRefPtr<BorderRadiusStyleValue> create(LengthPercentage const& horizontal_radius, LengthPercentage const& vertical_radius)
-    {
-        return adopt_ref(*new BorderRadiusStyleValue(horizontal_radius, vertical_radius));
-    }
-    virtual ~BorderRadiusStyleValue() override = default;
-
-    LengthPercentage const& horizontal_radius() const { return m_properties.horizontal_radius; }
-    LengthPercentage const& vertical_radius() const { return m_properties.vertical_radius; }
-    bool is_elliptical() const { return m_properties.is_elliptical; }
-
-    virtual ErrorOr<String> to_string() const override;
-
-    bool properties_equal(BorderRadiusStyleValue const& other) const { return m_properties == other.m_properties; }
-
-private:
-    BorderRadiusStyleValue(LengthPercentage const& horizontal_radius, LengthPercentage const& vertical_radius)
-        : StyleValueWithDefaultOperators(Type::BorderRadius)
-        , m_properties { .is_elliptical = horizontal_radius != vertical_radius, .horizontal_radius = horizontal_radius, .vertical_radius = vertical_radius }
-    {
-    }
-
-    virtual ValueComparingNonnullRefPtr<StyleValue const> absolutized(CSSPixelRect const& viewport_rect, Gfx::FontPixelMetrics const& font_metrics, CSSPixels font_size, CSSPixels root_font_size, CSSPixels line_height, CSSPixels root_line_height) const override;
-
-    struct Properties {
-        bool is_elliptical;
-        LengthPercentage horizontal_radius;
-        LengthPercentage vertical_radius;
-        bool operator==(Properties const&) const = default;
-    } m_properties;
 };
 
 class CalculatedStyleValue : public StyleValue {
